Datatype change

+++ benchmarks/src/cuda/gpuwattch-ubench/memories_benchmarks/BE_L1D_HIT/BE_L1D_HIT.cu
@@ -36,7 +36,7 @@ __global__ void l1_pointers_init(uint64_t *posArray){
   }
 }
 
-__global__ void l1_stress(uint64_t *posArray, uint64_t *dsink, uint64_t iterations){
+__global__ void l1_stress(uint64_t *posArray, uint64_t *dsink, unsigned iterations){
 
   // thread index
   uint32_t tid = blockIdx.x*blockDim.x + threadIdx.x;
@@ -59,7 +59,7 @@ __global__ void l1_stress(uint64_t *posArray, uint64_t *dsink, uint64_t iteratio
   	// pointer-chasing iterations times
   	// use ca modifier to cache the load in L1
   	#pragma unroll 100
-  	for(uint64_t i=0; i<iterations; ++i) { 
+  	for(unsigned i=0; i<iterations; ++i) { 
   	  asm volatile ("{\t\n"
   	    "ld.global.ca.u64 %0, [%1];\n\t"
   	    "}" : "=l"(ptr0) : "l"((uint64_t*)ptr1) : "memory"
@@ -74,13 +74,13 @@ __global__ void l1_stress(uint64_t *posArray, uint64_t *dsink, uint64_t iteratio
 }
 
 int main(int argc, char** argv){
-  uint64_t iterations;
+  unsigned iterations;
   if (argc != 2){
     fprintf(stderr,"usage: %s #iterations #cores #ActiveThreadsperWarp\n",argv[0]);
     exit(1);
   }
   else {
-    iterations = atoll(argv[1]);
+    iterations = atoi(argv[1]);
   }
   int total_threads = THREADS_NUM*NUM_BLOCKS;
  printf("Power Microbenchmarks with iterations %lu\n",iterations);
