+++ benchmarks/src/cuda/gpuwattch-ubench/static_power_modeling/INT_FP_TEX_STATIC/INT_FP_TEX_STATIC.cu
@@ -74,7 +74,7 @@ texture<float,1,cudaReadModeElementType> texmem7;
 texture<float,1,cudaReadModeElementType> texmem9;
 texture<float,1,cudaReadModeElementType> texmem8;
 
-__global__ void power_microbench(float *data1, float *data2, uint32_t *data3, uint32_t *data4, float *res, int div, uint64_t iterations, float* out, unsigned size) {
+__global__ void power_microbench(float *data1, float *data2, uint32_t *data3, uint32_t *data4, float *res, int div, unsigned iterations, float* out, unsigned size) {
 
   int gid = blockIdx.x*blockDim.x + threadIdx.x;
   register float s1 = data1[gid];
@@ -92,7 +92,7 @@ __global__ void power_microbench(float *data1, float *data2, uint32_t *data3, ui
     if((gid%32)<div){
     //ROI
       #pragma unroll 100
-      for (uint64_t j=0 ; j<iterations ; ++j) {
+      for (unsigned j=0 ; j<iterations ; ++j) {
         asm volatile ("{\t\n"
             "add.f32 %0, %1, %0;\n\t"
             "add.u32 %2, %3, %2;\n\t"
@@ -125,7 +125,7 @@ __global__ void power_microbench(float *data1, float *data2, uint32_t *data3, ui
 }
 
 int main(int argc, char** argv){
-  uint64_t iterations;
+  unsigned iterations;
   int blocks;
   int div;
   if (argc != 4){
@@ -133,12 +133,12 @@ int main(int argc, char** argv){
     exit(1);
   }
   else {
-    iterations = atoll(argv[1]);
+    iterations = atoi(argv[1]);
     blocks = atoi(argv[2]);
     div = atoi(argv[3]);
   }
  
- printf("Power Microbenchmarks with iterations %lu\n",iterations);
+ printf("Power Microbenchmarks with iterations %u\n",iterations);
