+++ benchmarks/src/cuda/gpuwattch-ubench/static_power_modeling/INT_STATIC/INT_STATIC.cu
@@ -16,7 +16,7 @@ inline void gpuAssert(cudaError_t code, const char *file, int line, bool abort=t
 
 
 template <class T>
-__global__ void power_microbench(T *data1, T *data2, T *res, int div, uint64_t iterations) {
+__global__ void power_microbench(T *data1, T *data2, T *res, int div, unsigned iterations) {
 
   int gid = blockIdx.x*blockDim.x + threadIdx.x;
   register T s1 = data1[gid];
@@ -31,7 +31,7 @@ __global__ void power_microbench(T *data1, T *data2, T *res, int div, uint64_t i
   if((gid%32)<div){
   //ROI
     #pragma unroll 100
-    for (uint64_t j=0 ; j<iterations ; ++j) {
+    for (unsigned j=0 ; j<iterations ; ++j) {
       asm volatile ("{\t\n"
           "add.u32 %0, %2, %0;\n\t"
           "add.u32 %0, %1, %0;\n\t"
@@ -57,7 +57,7 @@ __global__ void power_microbench(T *data1, T *data2, T *res, int div, uint64_t i
 }
 
 int main(int argc, char** argv){
-  uint64_t iterations;
+  unsigned iterations;
   int blocks;
   int div;
   if (argc != 4){
@@ -65,12 +65,12 @@ int main(int argc, char** argv){
     exit(1);
   }
   else {
-    iterations = atoll(argv[1]);
+    iterations = atoi(argv[1]);
     blocks = atoi(argv[2]);
     div = atoi(argv[3]);
   }
  
- printf("Power Microbenchmarks with iterations %lu\n",iterations);
+ printf("Power Microbenchmarks with iterations %u\n",iterations);
  int total_threads = THREADS_PER_BLOCK*blocks;
